requirements.txt
pyaudio
torch
torchaudio
openai-whisper
numpy
transformers
sentencepiece
sacremoses